requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-filter==2.0.0
get==2018.11.19
idna==2.8
post==2018.11.20
public==2018.11.20
pytz==2018.9
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
urllib3==1.24.1
